requirements.txt
#Discord Voice
py-cord[voice]

aio-pika
pyyaml

# Envinroment file .env
python-dotenv

# Report Generators
reportlab
markdown

# Lets keep our code clean
black
pylint

#openai and speech recognition
openai
faster_whisper
speechrecognition

# torch support of cuda 12
#--extra-index-url https://download.pytorch.org/whl/cu121
#torch

# torch cpu only support
--extra-index-url https://download.pytorch.org/whl/cpu
torch 

